requirements.txt
# A股量化交易系统 - 核心依赖
# 安装: pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple/

# 数据获取（免费无需注册）
akshare>=1.12.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# 技术指标（纯Python，无需编译C库）
pandas-ta>=0.3.14b

# Web界面
streamlit>=1.30.0
plotly>=5.17.0

# 定时任务
apscheduler>=3.10.0

# 工具
loguru>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
sqlalchemy>=2.0.0
mplfinance>=0.12.10b0
matplotlib>=3.7.0
